+++ freeman-core/src/Core/Plugin.php
@@ -14,7 +14,7 @@ defined( 'ABSPATH' ) || exit;
  */
 final class Plugin {
 
-	const VERSION = '1.10.13';
+	const VERSION = '1.10.16';
 
 	/**
 	 * Singleton instance.
@@ -156,6 +156,7 @@ final class Plugin {
 			$this->hub->boot();
 			( new \Freeman\Core\Admin\Dashboard( $this ) )->boot();
 			$this->importer->boot();
+			( new Settings_Tools() )->boot();
 		}
 
 		$this->booted = true;